+++ allauth/headless/urls.py
@@ -1,6 +1,7 @@
 from django.urls import include, path
 
 from allauth import app_settings as allauth_settings
+from allauth.headless import app_settings
 from allauth.headless.account import urls as account_urls
 from allauth.headless.base import urls as base_urls
 from allauth.headless.constants import Client
@@ -74,3 +75,13 @@ urlpatterns = [
         include((build_urlpatterns(Client.APP), "headless"), namespace="app"),
     ),
 ]
+
+if app_settings.SERVE_SPECIFICATION:
+    urlpatterns.append(
+        path(
+            "",
+            include(
+                "allauth.headless.spec.urls",
+            ),
+        )
+    )